requirements.txt
matrix-nio[e2ee]
google-generativeai
asyncio